små fixes (Account på bet, Field på bet, Balance på Account. Mere logik i Rouletten(Skulle gerne virke meget bedre nu). Seed ændres hvert 5. minut)

+++ CSGO_MVC/Models/Balance.cs
@@ -12,7 +12,7 @@ namespace CSGO_MVC.Models
         public int Id { get; set; }
         public double Amount { get; set; } //Bugger hvis begge hedder "Balance" muligvis wallet istedet for .. 
 
-        public Balance (int amount)
+        public Balance (double amount)
         {
             this.Amount = amount;
         }
